requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
confluent-kafka==2.4.0
idna==3.7
psycopg2-binary==2.9.9
requests==2.32.3
simplejson==3.19.2
urllib3==2.2.1
